initial simulator task that echoes the jtag console

+++ main/sensor/sensor_manager.c
@@ -15,40 +15,60 @@ bool barometer_data_ready = false;
 bool imu_data_ready = true; // todo: implement ISR for IMU
 bool gps_data_ready = true; // todo: implement ISR for GPS
 
+// FAKE SENSOR DATA VARIABLES (FOR SOFTWARE-IN-THE-LOOP)
+_Atomic barometer_sample_t sim_baro;
+_Atomic barometer_velocity_t sim_baro_vel;
+_Atomic acc_sample_t sim_low_g_acc;
+_Atomic acc_sample_t sim_high_g_acc;
+_Atomic gyr_sample_t sim_gyr;
+_Atomic gps_sample_t sim_gps;
+
 // MARK: SENSOR INITIALIZATION
-esp_err_t initialize_sensors(void)
+esp_err_t initialize_sensors(bool is_simulator)
 {
-    esp_err_t ret;
-    i2c_port_t BMP390_I2C_PORT = R3_I2C1_PORT;
-    i2c_port_t SAM_M10Q_I2C_PORT = R3_I2C0_PORT;
-    spi_host_device_t LSM_SPI_HOST = SPI3_HOST;
+    if(is_simulator) {
+        simulator = true;
+        barometer_data_ready = true;
+        imu_data_ready = true;
+        gps_data_ready = true;
+    } else {
+        esp_err_t ret;
+        i2c_port_t BMP390_I2C_PORT = R3_I2C1_PORT;
+        i2c_port_t SAM_M10Q_I2C_PORT = R3_I2C0_PORT;
+        spi_host_device_t LSM_SPI_HOST = SPI3_HOST;
 
-    // Install ISR service
-    ret = gpio_install_isr_service(0);
-    if (ret != ESP_OK) {
-        ESP_LOGE(TAG, "Failed to install ISR service");
-        return ret;
+        // Install ISR service
+        ret = gpio_install_isr_service(0);
+        if (ret != ESP_OK) {
+            ESP_LOGE(TAG, "Failed to install ISR service");
+            return ret;
+        }
+
+        ret = bmp390_flight_init(BMP390_I2C_PORT, barometer_int_callback, &bmp_ctx);
+        if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE BMP390"); return ret;}
+
+        vTaskDelay(pdMS_TO_TICKS(1000));
+
+        ret = GPS_init(SAM_M10Q_I2C_PORT);
+        if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE SAM-M10Q"); return ret;}
+
+        ret = lsm_flight_init(LSM_SPI_HOST);
+        if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE LSM6DSV320X"); return ret;}
+
+        ESP_LOGI(TAG, "SUCCESSFULLY INITIALIZED ALL SENSORS");
     }
-
-    ret = bmp390_flight_init(BMP390_I2C_PORT, barometer_int_callback, &bmp_ctx);
-    if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE BMP390"); return ret;}
-
-    vTaskDelay(pdMS_TO_TICKS(1000));
-    
-    ret = GPS_init(SAM_M10Q_I2C_PORT);
-    if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE SAM-M10Q"); return ret;}
-
-    ret = lsm_flight_init(LSM_SPI_HOST);
-    if(ret != ESP_OK) {ESP_LOGE(TAG, "FAILED TO INITIALIZE LSM6DSV320X"); return ret;}
-
-    ESP_LOGI(TAG, "SUCCESSFULLY INITIALIZED ALL SENSORS");
     return ESP_OK;
 }
 
 void poll_baro(barometer_sample_t *pBaro)
 {
     if(simulator) {
-        ;
+        barometer_sample_t temp_baro = atomic_load(&sim_baro);
+        pBaro->timestamp = esp_timer_get_time();
+        pBaro->pressure = temp_baro.pressure;
+        pBaro->temperature = temp_baro.temperature;
+        pBaro->altitude_agl = temp_baro.altitude_agl;
+        pBaro->ground_altitude = temp_baro.ground_altitude;
     }
     else { 
         baro_double_t bmp390_out;
@@ -70,7 +90,15 @@ void barometer_int_callback(void *args)
 esp_err_t poll_gps(gps_sample_t *gps)
 {
     if(simulator) {
-        ;
+        gps_sample_t temp_sam_m10q_data = atomic_load(&sim_gps);
+        gps->timestamp = esp_timer_get_time();
+        gps->UTCtstamp = temp_sam_m10q_data.UTCtstamp;
+        gps->lat = temp_sam_m10q_data.lat;
+        gps->lon = temp_sam_m10q_data.lon;
+        gps->altitude_ellipsoid = temp_sam_m10q_data.altitude_ellipsoid;
+        gps->altitude_msl = temp_sam_m10q_data.altitude_msl;
+        gps->fixType = temp_sam_m10q_data.fixType;
+        gps->num_sats = temp_sam_m10q_data.num_sats;
     }
     else {
         GPS_data_t sam_m10q_data;
@@ -90,25 +118,49 @@ esp_err_t poll_gps(gps_sample_t *gps)
 
 static esp_err_t poll_imu(acc_sample_t *high_g, acc_sample_t *low_g, gyr_sample_t *gyr)
 {
-    lsm_raw_data_t raw_imu_data;
-    esp_err_t ret = lsm_get_local(&raw_imu_data);
-    if(ret) return ret;
-    int64_t timestamp = esp_timer_get_time();
+    if(simulator) {
+        acc_sample_t temp_low_acc_data = atomic_load(&sim_low_g_acc);
+        acc_sample_t temp_high_acc_data = atomic_load(&sim_high_g_acc);
+        gyr_sample_t temp_gyr_sample = atomic_load(&sim_gyr);
+        int64_t timestamp = esp_timer_get_time();
+        
+        low_g->timestamp = timestamp;
+        low_g->acc_x = temp_low_acc_data.acc_x;
+        low_g->acc_y = temp_low_acc_data.acc_y;
+        low_g->acc_z = temp_low_acc_data.acc_z;
 
-    high_g->timestamp = timestamp;
-    high_g->acc_x = raw_imu_data.highacc_x;
-    high_g->acc_y = raw_imu_data.highacc_y;
-    high_g->acc_z = raw_imu_data.highacc_z;
+        high_g->timestamp = timestamp;
+        high_g->acc_x = temp_high_acc_data.acc_x;
+        high_g->acc_y = temp_high_acc_data.acc_y;
+        high_g->acc_z = temp_high_acc_data.acc_z;
 
-    low_g->timestamp = timestamp;
-    low_g->acc_x = raw_imu_data.lowacc_x;
-    low_g->acc_y = raw_imu_data.lowacc_y;
-    low_g->acc_z = raw_imu_data.lowacc_z;
+        gyr->timestamp = timestamp;
+        gyr->gyr_x = temp_gyr_sample.gyr_x;
+        gyr->gyr_y = temp_gyr_sample.gyr_y;
+        gyr->gyr_z = temp_gyr_sample.gyr_z;
+    }
+    else {
+        lsm_raw_data_t raw_imu_data;
+        esp_err_t ret = lsm_get_local(&raw_imu_data);
+        if(ret) return ret;
+        int64_t timestamp = esp_timer_get_time();
 
-    gyr->timestamp = timestamp;
-    gyr->gyr_x = raw_imu_data.gyr_x;
-    gyr->gyr_y = raw_imu_data.gyr_y;
-    gyr->gyr_z = raw_imu_data.gyr_z;
+        high_g->timestamp = timestamp;
+        high_g->acc_x = raw_imu_data.highacc_x;
+        high_g->acc_y = raw_imu_data.highacc_y;
+        high_g->acc_z = raw_imu_data.highacc_z;
+
+        low_g->timestamp = timestamp;
+        low_g->acc_x = raw_imu_data.lowacc_x;
+        low_g->acc_y = raw_imu_data.lowacc_y;
+        low_g->acc_z = raw_imu_data.lowacc_z;
+
+        gyr->timestamp = timestamp;
+        gyr->gyr_x = raw_imu_data.gyr_x;
+        gyr->gyr_y = raw_imu_data.gyr_y;
+        gyr->gyr_z = raw_imu_data.gyr_z;
+    }
+    
 
     return ESP_OK;
 }
@@ -175,4 +227,15 @@ void poll_sensors(barometer_sample_t *pBaro, barometer_velocity_t *pBaro_vel, ac
     if (gps_data_ready){
         // poll_gps(gps);
     }
+}
+
+void feed_fake_flight_data(barometer_sample_t baro, barometer_velocity_t baro_vel, acc_sample_t high_g, acc_sample_t low_g, gyr_sample_t gyr, gps_sample_t gps)
+{
+    // Atomic store all these variables
+    atomic_store(&sim_baro, baro);
+    atomic_store(&sim_baro_vel, baro_vel);
+    atomic_store(&sim_high_g_acc, high_g);
+    atomic_store(&sim_low_g_acc, low_g);
+    atomic_store(&sim_gyr, gyr);
+    atomic_store(&sim_gps, gps);
 }
